Remove unused dependency
The falafel dependency was used exclusively in a previous iteration of
this module to programmatically alter moment to stub the date. See:
https://github.com/goodeggs/geomoment/blob/e2e123938996b2687a472f59348894712eb0e2ff/util/patch_moment.coffee

That functionality has subsequently been removed and nothing is requiring
falafel at present. Unfortunately, falafel at this particular version
(0.31) specified esprima as a '*' version dependency, and I somehow downloaded
an invalid version of esprima which made npm very sad. Fortunately, I can just
remove falafel now that it is unused :rocket: